Reworked Shoot + Napalm
- Reorganized the Shoot function to instead return a Projectile
- This was done in order for projectile properties to be properly applied to applicable projectiles
- Added 'Napalm'. A hardmode material that allows weapon to inflict 'Napalmed' scorching enemies.
- Napalm does extra damage to spiders and gel based enemies

+++ Core/Forge/ForgePlayer.cs
@@ -16,12 +16,12 @@ namespace Disarray.Core.Forge
 
         public override void ModifyHitNPC(Item item, NPC target, ref int damage, ref float knockback, ref bool crit)
         {
-            damage = (int)((float)damage * Damage) + DamageFlat;
+            damage = (int)(damage * Damage) + DamageFlat;
         }
 
         public override void ModifyHitNPCWithProj(Projectile proj, NPC target, ref int damage, ref float knockback, ref bool crit, ref int hitDirection)
         {
-            damage = (int)((float)damage * Damage) + DamageFlat;
+            damage = (int)(damage * Damage) + DamageFlat;
         }
     }
 }
